backend/requirements.txt
# FastAPI backend requirements
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-dotenv
websockets
